Makefile: Make it possible to build bl301.bin on demand
Previously the build steps for bl301.bin were disabled completely due to
a problem with Yocto builds. This change makes it possible to rebuild that
binary with

    make bl301.bin

However, the precompiled binary remains in the repository and bl301.bin is
not built per default. Also "make distclean" doesn't delete the binary,
just as before.

Signed-off-by: Alexander F. Mayer <alexander.mayer@streamunlimited.com>
Acked-by: Martin Pietryka <martin.pietryka@streamunlimited.com>