board: sue: stream1832: remove unused Amlogic code
CONFIG_AML_TINY_USBTOOL is not set in our config, thus upgrade_step handling
can be safely removed. I'm not 100% sure about the reboo_mode handling,
it might actually create issues with the Amlogic flash-tool.

However both of those variables relied on saveenv which we do not want to
use, because we do not want to have the full env in the environment partition.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>