HACK: usb: v2_burning: optimus_transform: filter save and saveenv bulkcmds
We want to prevent the Amlogic flash-tool to run save or saveenv, because
that would write the full env into the enviroment partition. However our
programming flow requires only changed env variables int the NAND, the rest
of the env should be taken from the defaults. Thus we just ignore those calls
from the Amlogic flash-tool. However if someone really wants to save the
full environment to the NAND he can do that by using the saveenv_s bulkcmd
instead.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>