tools: env: do not load/store the default environment if it is empty
We only want to store version information in the env partition and not
the whole environment, however the userspace U-Boot tools will load,
display and write back the default env if no existing env was on the
partition. We want to avoid that behavior so that a module should never
have the environment saved in the flash. For the case that someone needs
to display/store the default environment an additional -d (--default)
flag was added, this flag will will load the default environment
(for both fw_printenv and fw_setenv) when the CRC (e.g. NAND partition
was empty) is wrong.

One downside with this is that if there is something in the environment,
it is not possible to print the default with this flag. However there is
no use case for this behavior and the help of the flag mentions this quirk.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>