board: streamunlimited: add support for the Stream195x module
Many of the common parts, e.g. ADC detection will not be really required
on the Stream195x. However these parts of the code were imported from
Stream1832 U-Boot, so we keep the for consistency.

The imported code comes from the `u-boot-stream183x_2015.01-r16` tag in
the `Stream800-uboot` repository. Commit id:
 * a73ee46f66cb9d3f800b851d6f93a0f12dd6cda1

The 1x4Gb timings where generated by Emil using the Excel sheet:
 * _u1__Stream1955_1xKingston4Gbit_800MHz_MX8M_Mini_DDR3L_RPA_v7.xlsx

The output of these Excel sheets is a `*.ds` file which together with the
`MX8M_DDR_tool` will generate the respective `*.c` files.

NOTE: The 2x2Gb timings originate from the following Excel sheet:
 * !!!_Stream1955_2xNanya2Gbit_MX8MM_845_DDR3L_register_programming_aid_VAL_preliminiary_v1.xlsx

But because in the future we will not support the 2x2Gb configuration,
these timings were not updated to the most recent ones by Emil.

We also make sure to disable fastboot on locked mofules for the main
reason of having another layer of security. The only reason fastboot
support is compiled in the first place it to allow for a faster loading
of the fitImage in the facotry, after that it should not be used for
anything else.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[rename StreamKit 3D to StreamKit or just kit]
[mark StreamKit prime as legacy]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>