fastboot: add FASTBOOT_SUE_FACTORY_SUBSET option
This option will only enable three fastboot commands. UCmd, ACmd and download.
These commands are enough to be able to boot the image in the factory over
fastboot. All other fastboot commands are disabled. Main reason being is that
the code is a mess and takes up space (a brief smoke check showed that the
linker is correctly removing unused cb_* functions).

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>