lib: rsa: Kconfig: don't use RSA_FREESCALE_EXP on i.MX8M and i.MX8
There was a previous commit excluding this feature for i.MX5/6/7, quoting:
> The CAAM in IMX parts doesn't support public key hardware acceleration
> (PKHA), so don't use RSA_FREESCALE_EXP.
> ...

I assume this is also the case for the i.MX8 because the RSA verification
of fit-Images was failing when RSA_FREESCALE_EXP was enabled.

NOTE: From the Kconfig definitions it seems that i.MX8 and i.MX8M are two
separate "architectures" (e.g. ARCH_IMX8M does not select ARCH_IMX8).

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>