board: sue: stream1832: set initial CPU frequency to 672 MHz [STREAMEIGHT-3745]
We need to set this voltage because if the CPU was reset but the AXP
not, e.g. warm reset the DCDC2 voltage might be at 0.85 V from when
the Linux governor set a low operating point before an external reset
and only 672MHz frequency is compatible with such a low voltage.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[written commit message body]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>